fixing identifier

+++ binning-dao/binning-dao-api/src/main/java/org/renci/binning/dao/clinbin/PhenotypeXDAO.java
@@ -3,6 +3,6 @@ package org.renci.binning.dao.clinbin;
 import org.renci.binning.dao.BaseDAO;
 import org.renci.binning.dao.clinbin.model.PhenotypeX;
 
-public interface PhenotypeXDAO extends BaseDAO<PhenotypeX, Long> {
+public interface PhenotypeXDAO extends BaseDAO<PhenotypeX, Integer> {
 
 }
